+++ src/main/java/io/zerogone/blog/post/comment/converter/CommentEntityToDtoConverter.java
@@ -1,10 +1,10 @@
-package io.zerogone.converter;
+package io.zerogone.blog.post.comment.converter;
 
-import io.zerogone.model.dto.CommentDto;
-import io.zerogone.model.dto.PostDto;
+import io.zerogone.blog.post.comment.model.CommentDto;
+import io.zerogone.blog.post.model.PostDto;
 import io.zerogone.user.model.UserDto;
-import io.zerogone.model.entity.Comment;
-import io.zerogone.model.entity.Post;
+import io.zerogone.blog.post.comment.model.Comment;
+import io.zerogone.blog.post.model.Post;
 import io.zerogone.user.model.User;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
