+++ src/main/java/io/zerogone/blog/post/controller/PostCreateController.java
@@ -1,8 +1,8 @@
-package io.zerogone.controller.api;
+package io.zerogone.blog.post.controller;
 
 import io.zerogone.exception.NotAuthorizedException;
 import io.zerogone.blog.model.BlogDto;
-import io.zerogone.model.dto.PostDto;
+import io.zerogone.blog.post.model.PostDto;
 import io.zerogone.user.model.UserDto;
 import io.zerogone.service.create.CreateService;
 import io.zerogone.service.search.SearchService;
