+++ src/main/java/io/zerogone/blog/post/controller/PostListController.java
@@ -1,7 +1,7 @@
-package io.zerogone.controller;
+package io.zerogone.blog.post.controller;
 
 import io.zerogone.blog.model.BlogName;
-import io.zerogone.model.dto.PostDto;
+import io.zerogone.blog.post.model.PostDto;
 import io.zerogone.service.search.SearchService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
